# core-common-customer-mgmt-sdk/src/main/resources/api-spec/openapi.yml
{
  "openapi": "3.0.1",
  "info": {
    "title": "core-common-customer-mgmt",
    "description": "Customer Management Core Application",
    "contact": {
      "name": "Firefly Software Foundation",
      "email": "dev@getfirefly.io"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://core.getfirefly.io/customers",
      "description": "Development Environment"
    },
    {
      "url": "/",
      "description": "Local Development Environment"
    }
  ],
...
  "paths": {
...
    "/api/v1/parties/{partyId}/natural-persons/{naturalPersonId}": {
      "get": {
        "tags": [
          "Natural Persons"
        ],
        "summary": "Get natural person by ID",
        "description": "Retrieve a specific natural person associated with a party by its unique identifier",
        "operationId": "getNaturalPersonById",
        "parameters": [
          {
            "name": "partyId",
            "in": "path",
            "description": "Unique identifier of the party",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "naturalPersonId",
            "in": "path",
            "description": "Unique identifier of the natural person",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "description": "Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Natural person found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/NaturalPersonDTO"
                }
              }
            }
          },
          "404": {
            "description": "Natural person or party not found"
          }
        }
      },
      "put": {
        "tags": [
          "Natural Persons"
        ],
        "summary": "Update natural person",
        "description": "Update an existing natural person associated with a party",
        "operationId": "updateNaturalPerson",
        "parameters": [
          {
            "name": "partyId",
            "in": "path",
            "description": "Unique identifier of the party",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "naturalPersonId",
            "in": "path",
            "description": "Unique identifier of the natural person",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "description": "Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NaturalPersonDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Invalid natural person data"
          },
          "200": {
            "description": "Natural person successfully updated",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/NaturalPersonDTO"
                }
              }
            }
          },
          "404": {
            "description": "Natural person or party not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Natural Persons"
        ],
        "summary": "Delete natural person",
        "description": "Delete a natural person associated with a party from the system",
        "operationId": "deleteNaturalPerson",
        "parameters": [
          {
            "name": "partyId",
            "in": "path",
            "description": "Unique identifier of the party",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "naturalPersonId",
            "in": "path",
            "description": "Unique identifier of the natural person",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "description": "Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Natural person or party not found"
          },
          "204": {
            "description": "Natural person successfully deleted"
          }
        }
      }
    },
...
    "/api/v1/parties/{partyId}/natural-persons": {
      "get": {
        "tags": [
          "Natural Persons"
        ],
        "summary": "Get natural person by party ID",
        "description": "Retrieve the natural person associated with a specific party",
        "operationId": "getNaturalPersonByPartyId",
        "parameters": [
          {
            "name": "partyId",
            "in": "path",
            "description": "Unique identifier of the party",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "description": "Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Party not found"
          },
          "200": {
            "description": "Natural person found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/NaturalPersonDTO"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Natural Persons"
        ],
        "summary": "Create natural person for a party",
        "description": "Create a new natural person associated with a specific party",
        "operationId": "createNaturalPerson",
        "parameters": [
          {
            "name": "partyId",
            "in": "path",
            "description": "Unique identifier of the party",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "description": "Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NaturalPersonDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Natural person successfully created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/NaturalPersonDTO"
                }
              }
            }
          },
          "400": {
            "description": "Invalid natural person data"
          },
          "404": {
            "description": "Party not found"
          }
        }
      }
    },
    "/api/v1/parties/{partyId}/natural-persons/filter": {
      "post": {
        "tags": [
          "Natural Persons"
        ],
        "summary": "Filter natural persons for a party",
        "description": "Retrieve a paginated list of natural persons associated with a specific party based on filtering criteria",
        "operationId": "filterNaturalPersons",
        "parameters": [
          {
            "name": "partyId",
            "in": "path",
            "description": "Unique identifier of the party",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "description": "Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FilterRequestNaturalPersonDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successfully retrieved filtered natural persons",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PaginationResponse"
                }
              }
            }
          },
          "404": {
            "description": "Party not found"
          },
          "400": {
            "description": "Invalid filter request"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "NaturalPersonDTO": {
        "required": [
          "familyName1",
          "givenName",
          "partyId"
        ],
        "type": "object",
        "properties": {
          "naturalPersonId": {
            "type": "string",
            "format": "uuid",
            "readOnly": true
          },
          "partyId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "givenName": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "middleName": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "familyName1": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "familyName2": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date"
          },
          "birthPlace": {
            "maxLength": 150,
            "minLength": 0,
            "type": "string"
          },
          "birthCountryId": {
            "type": "string",
            "format": "uuid"
          },
          "nationalityCountryId": {
            "type": "string",
            "format": "uuid"
          },
          "gender": {
            "type": "string",
            "enum": [
              "MALE",
              "FEMALE",
              "NON_BINARY",
              "UNSPECIFIED"
            ]
          },
          "maritalStatus": {
            "type": "string",
            "enum": [
              "SINGLE",
              "MARRIED",
              "DIVORCED",
              "WIDOWED"
            ]
          },
          "taxIdNumber": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "residencyStatus": {
            "type": "string",
            "enum": [
              "RESIDENT",
              "NON_RESIDENT"
            ]
          },
          "occupation": {
            "maxLength": 150,
            "minLength": 0,
            "type": "string"
          },
          "monthlyIncome": {
            "minimum": 0,
            "exclusiveMinimum": true,
            "type": "number"
          },
          "suffix": {
            "maxLength": 20,
            "minLength": 0,
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "FilterOptions": {
        "type": "object",
        "properties": {
          "caseInsensitiveStrings": {
            "type": "boolean"
          },
          "includeInheritedFields": {
            "type": "boolean"
          }
        },
        "description": "Filter options for customizing filter behavior"
      },
...
      "PaginationRequest": {
        "type": "object",
        "properties": {
          "pageNumber": {
            "type": "integer",
            "description": "The zero-based page number to retrieve.",
            "format": "int32",
            "example": 0,
            "default": 0
          },
          "pageSize": {
            "type": "integer",
            "description": "The number of items per page.",
            "format": "int32",
            "example": 10,
            "default": 10
          },
          "sortBy": {
            "type": "string",
            "description": "The field to sort the results by.",
            "example": "name"
          },
          "sortDirection": {
            "type": "string",
            "description": "The direction of sorting, either ASC or DESC.",
            "example": "DESC",
            "default": "DESC"
          }
        },
        "description": "Represents a pagination request for retrieving paginated results, including page number, size, sort field, and direction."
      },
      "RangeFilter": {
        "type": "object",
        "properties": {
          "ranges": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/RangeObject"
            },
            "description": "Map of field names to their range criteria"
          }
        },
        "description": "Represents range filters for numeric fields"
      },
      "RangeObject": {
        "type": "object",
        "properties": {
          "from": {
            "type": "object",
            "description": "Start value of the range"
          },
          "to": {
            "type": "object",
            "description": "End value of the range"
          }
        },
        "description": "Represents a range with from and to values"
      },
      "PaginationResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "description": "The list of items for the current page.",
            "items": {
              "type": "object",
              "description": "The list of items for the current page."
            }
          },
          "totalElements": {
            "type": "integer",
            "description": "The total number of elements across all pages.",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "description": "The total number of pages based on the data size and page size.",
            "format": "int32"
          },
          "currentPage": {
            "type": "integer",
            "description": "The current page number, typically zero-based.",
            "format": "int32"
          }
        },
        "description": "Represents a paginated response containing a list of items and pagination metadata."
      },
...
      "FilterRequestNaturalPersonDTO": {
        "required": [
          "pagination"
        ],
        "type": "object",
        "properties": {
          "filters": {
            "$ref": "#/components/schemas/NaturalPersonDTO"
          },
          "rangeFilters": {
            "$ref": "#/components/schemas/RangeFilter"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationRequest"
          },
          "options": {
            "$ref": "#/components/schemas/FilterOptions"
          }
        },
        "description": "Filter criteria for natural persons"
      },
...
    }
  }
}